Further tests refactor
